+++ CHANGELOG.md
@@ -6,11 +6,13 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Accessibility: Color blind mode
 - Accessibility: Change text size
+- map change for each level
 
 ### Fixed
 - lock level until the player finish the one before
 - drag and drop on the map, other than the tile or red tile
 - when a level end, the next level couldn't load the chapter
+- tree in front of the camera is now blur
 
 
 ## [0.1.5-alpha] - 2026-07-11